refactor: rename update-user folder to register-user
