+++ src/main/factories/usecases/user/update-user-usecase-factory.ts
@@ -1,5 +1,5 @@
 import { type UpdateUser } from '@/domain/contracts/user'
-import { UpdateUserUseCase } from '@/usecases/implementations/user/update-user/register-user-usecase'
+import { UpdateUserUseCase } from '@/usecases/implementations/user/register-user/register-user-usecase'
 import { makeUpdateUserPrismaRepo } from '../../infra/db/prisma/user/update-user-prisma-repo-factory'
 
 export const makeUpdateUserUseCase = (): UpdateUser => {
